feat!(economy): Update schema types
