+++ commands/economy/economy.go
@@ -3,21 +3,22 @@ package economy
 //go:generate sqlboiler --no-hooks psql
 
 import (
-	"github.com/RhykerWells/asbwig/commands/economy/balance"
+	"github.com/RhykerWells/asbwig/commands/economy/informational/balance"
+	set "github.com/RhykerWells/asbwig/commands/economy/informational/settings"
 	"github.com/RhykerWells/asbwig/common"
 	"github.com/RhykerWells/asbwig/common/dcommand"
 )
 
 var GuildEconomySchema = []string{`
 CREATE TABLE IF NOT EXISTS economy_config (
-	guild_id BIGINT PRIMARY KEY,
+	guild_id TEXT PRIMARY KEY,
 	max_bet BIGINT NOT NULL DEFAULT '5000',
 	symbol TEXT NOT NULL DEFAULT '£',
 	start_balance BIGINT NOT NULL DEFAULT '200'
 );
 `,`
 CREATE TABLE IF NOT EXISTS economy_cash (
-	guild_id BIGINT PRIMARY KEY,
+	guild_id TEXT PRIMARY KEY,
 	user_id BIGINT NOT NULL,
 	cash BIGINT
 )
@@ -28,6 +29,7 @@ func EconomySetup(cmdHandler *dcommand.CommandHandler) {
 
 	cmdHandler.RegisterCommands(
 		balance.Command,
+		set.Command,
 	)
 }
